requirements.txt
pandas>=2.0
PyYAML>=6.0
numpy>=1.26.0
jsonschema>=4.17.0
